--- a/p4_symbolic/symbolic/expected/conditional_nonlattice.txt
+++ b/p4_symbolic/symbolic/expected/conditional_nonlattice.txt
@@ -1,7 +1,7 @@
 ================================================================================
 Finding packet for table MyIngress.t_1 and row -1
 ================================================================================
-ingress_port = #b000001001
+ingress_port = #b000010001
 egress_port = #b000000001
 trace:
 dropped = 0
@@ -33,7 +33,7 @@ standard_metadata.egress_spec = #b000000000
 standard_metadata.enq_qdepth = #b0000000000000000000
 standard_metadata.enq_timestamp = #x00000000
 standard_metadata.ingress_global_timestamp = #x000000000000
-standard_metadata.ingress_port = #b000001001
+standard_metadata.ingress_port = #b000010001
 standard_metadata.instance_type = #x00000000
 standard_metadata.mcast_grp = #x0000
 standard_metadata.packet_length = #x00000000
@@ -63,7 +63,7 @@ standard_metadata.egress_spec = #b000000000
 standard_metadata.enq_qdepth = #b0000000000000000000
 standard_metadata.enq_timestamp = #x00000000
 standard_metadata.ingress_global_timestamp = #x000000000000
-standard_metadata.ingress_port = #b000001001
+standard_metadata.ingress_port = #b000010001
 standard_metadata.instance_type = #x00000000
 standard_metadata.mcast_grp = #x0000
 standard_metadata.packet_length = #x00000000
@@ -93,7 +93,7 @@ standard_metadata.egress_spec = #b000000001
 standard_metadata.enq_qdepth = #b0000000000000000000
 standard_metadata.enq_timestamp = #x00000000
 standard_metadata.ingress_global_timestamp = #x000000000000
-standard_metadata.ingress_port = #b000001001
+standard_metadata.ingress_port = #b000010001
 standard_metadata.instance_type = #x00000000
 standard_metadata.mcast_grp = #x0000
 standard_metadata.packet_length = #x00000000